Address feedback: move files to auxilary section + add gate_set input
