--- src/braket/experimental/algorithms/random_circuit/random_circuit.md
@@ -1,7 +1,0 @@
-The `random_circuit` function generates a random quantum circuit using the Amazon Braket SDK. It's designed to construct a circuit with a specified number of qubits, instructions, and a maximum  number of qubits the gate acts on. This function is useful for creating diverse quantum circuits for testing quantum algorithms or for benchmarking gate-based QPUs.
-
-<!--
-[metadata-name]: Random Quantum Circuit Generator
-[metadata-tags]: Textbook
-[metadata-url]: https://github.com/aws-samples/amazon-braket-algorithm-library/tree/main/src/braket/experimental/algorithms/random_circuit
--->
