+++ src/braket/experimental/auxilary_functions/random_circuit/random_circuit.md
@@ -1,0 +1,7 @@
+The `random_circuit` function generates a random quantum circuit using the Amazon Braket SDK.
+The function creates a diverse set of circuits that can be used for testing Braket SDK functionality and compiler benchmarking.
+<!--
+[metadata-name]: Random Quantum Circuit Generator
+[metadata-tags]: Textbook
+[metadata-url]: https://github.com/aws-samples/amazon-braket-algorithm-library/tree/main/src/braket/experimental/algorithms/random_circuit
+-->
